2.0.0 - MDCL is now a Compiler!
Yup. MDCL is now a Compiler. This is a continuation of MDCL-Interpreter, but as a Compiler.
The MDCL Compiler currently compiles into Python, but in the future I'm hoping to be able to compile MDCL into other languages, specifically low-level ones for extra execution speed.

So yeah, here it is.